refactor(swiper): 提取滑动位置计算逻辑到独立函数
- 将 modulus 和 getSwipeToPosition 提取为公共函数
- 更新测试用例使用新函数
